# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: selfmx API
  description: |
    Self-hosted Email the easy way. Enterprise-grade user management,
    multi-provider notification system, and multi-LLM orchestration.
  version: 0.1.0
  contact:
    name: selfmx Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
...
paths:
...
  /email-templates:
    get:
      tags: [Email Templates]
      summary: List email templates
      security:
        - sanctum: []
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EmailTemplate'
  /email-templates/{key}:
    get:
      tags: [Email Templates]
      summary: Get email template
      security:
        - sanctum: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template details
    put:
      tags: [Email Templates]
      summary: Update email template
      security:
        - sanctum: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: string
                body_html:
                  type: string
                body_text:
                  type: string
      responses:
        '200':
          description: Template updated
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      description: Laravel Sanctum session-based authentication
...